Arreglamos problema de dropzone al arrastrar varias imagenes simultaneamente

+++ resources/js/dropzone.js
@@ -10,13 +10,14 @@ export const dropzone = new Dropzone('#dropzone', {
 	maxFiles: 1,
 	maxFilesize: 3,
 	uploadMultiple: false,
-	
+
 	init: function () {
 		this.on('addedfile', function (file) {
-			// Si se agrega una segunda imagen, remueve la imagen anterior
-			if (this.files.length > 1) {
-				this.removeFile(this.files[0]);
-			}
+			if (this.files.length > 1) this.removeFile(this.files[0]);
+		});
+		this.on('maxfilesexceeded', function (file) {
+			this.removeAllFiles();
+			this.addFile(file);
 		});
 	},
 });
